+++ src/paco/application/reseng_iotpolicy.py
@@ -1,12 +1,11 @@
-import paco.cftemplates
 from paco.application.res_engine import ResourceEngine
+from paco.stack.botostacks.iotpolicy import IoTPolicyBotoStack
 
 class IoTPolicyResourceEngine(ResourceEngine):
 
     def init_resource(self):
-        iotpolicy_ctl = self.paco_ctx.get_controller('IoTPolicy')
-        iotpolicy_ctl.add_iotpolicy(
-            self.account_ctx,
+        self.stack_group.add_new_boto_stack(
             self.aws_region,
-            self.resource
+            self.resource,
+            IoTPolicyBotoStack,
         )